// frontend_machine_learning/src/interfaces.ts
export interface UserData {
  Gender: string | null | null;
  Age: number | null;
  Height: number | null;
  Weight: number | null;
  family_history_with_overweight: string | null;
  FAVC: string | null;
  FCVC: number | null;
  NCP: number | null;
  CAEC: string | null;
  SMOKE: string | null;
  CH2O: number | null;
  SCC: string | null;
  FAF: number | null;
  TUE: number | null;
  CALC: string | null;
  MTRANS: string | null;
}